+++ tests/engine/test_engine_wrapper/dummy_engine.py
@@ -5,6 +5,7 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
+from flepimop2.configuration import IdentifierString
 from flepimop2.system.abc import SystemProtocol
 
 
@@ -12,7 +13,7 @@ def runner(
     f: SystemProtocol,
     times: NDArray[np.float64],
     state: NDArray[np.float64],
-    params: dict[str, Any],
+    params: dict[IdentifierString, Any],
     *,
     accumulate: bool,
 ) -> NDArray[np.float64]: